- Add the new BSON temporary type for testing   modified:   storage/connect/CMakeLists.txt   modified:   storage/connect/bson.cpp   modified:   storage/connect/bson.h   modified:   storage/connect/bsonudf.cpp   modified:   storage/connect/bsonudf.h   modified:   storage/connect/global.h   modified:   storage/connect/json.cpp   modified:   storage/connect/jsonudf.cpp   modified:   storage/connect/mysql-test/connect/disabled.def   modified:   storage/connect/mysql-test/connect/t/mongo_test.inc   modified:   storage/connect/plugutil.cpp   modified:   storage/connect/tabbson.cpp   modified:   storage/connect/tabjson.cpp

+++ storage/connect/global.h
@@ -185,7 +185,7 @@ typedef struct _global {            /* Global structure                */
   size_t    Sarea_Size;             /* Work area size                  */
 	PACTIVITY Activityp;
   char      Message[MAX_STR];				/* Message (result, error, trace)  */
-	ulong     More;										/* Used by jsonudf                 */
+	size_t    More;										/* Used by jsonudf                 */
 	size_t    Saved_Size;             /* Saved work area to_free         */
 	bool      Createas;               /* To pass multi to ext tables     */
   void     *Xchk;                   /* indexes in create/alter         */
